requirements.txt
redis==3.2.1
psycopg2-binary==2.7.7